+++ noodles-util/examples/util_variant_query.rs
@@ -18,7 +18,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let src = args.next().expect("missing src");
     let region = args.next().expect("missing region").parse()?;
 
-    let mut reader = variant::indexed_reader::Builder::default().build_from_path(src)?;
+    let mut reader = variant::io::indexed_reader::Builder::default().build_from_path(src)?;
     let header = reader.read_header()?;
 
     let query = reader.query(&header, &region)?;