+++ noodles-util/src/variant.rs
@@ -1,12 +1,6 @@
-//! I/O for variant formats.
+//! Variant format utilities.
 
-mod compression_method;
-mod format;
-pub mod indexed_reader;
-pub mod reader;
+pub mod io;
 pub mod writer;
 
-pub use self::{
-    compression_method::CompressionMethod, format::Format, indexed_reader::IndexedReader,
-    reader::Reader, writer::Writer,
-};
+pub use self::writer::Writer;